configure and add Redux with TS to application
