+++ src/store/reducers/UserSlice.ts
@@ -1,0 +1,29 @@
+import { createSlice } from '@reduxjs/toolkit'
+import type { PayloadAction } from '@reduxjs/toolkit'
+import type { RootState } from '../store'
+
+
+interface CounterState {
+  value: number
+}
+
+
+const initialState: CounterState = {
+  // @ts-ignore
+  value:0,
+}
+
+export const counterSlice = createSlice({
+  name: 'counter',
+
+  initialState,
+  reducers: {
+    increment: (state) => {
+      state.value += 1
+    },
+  },
+})
+
+export const { increment } = counterSlice.actions
+
+export default counterSlice.reducer
