+++ src/store/store.ts
@@ -1,0 +1,18 @@
+import { combineReducers, configureStore } from '@reduxjs/toolkit'
+import userReducer from './reducers/UserSlice'
+// ...
+
+const rootReducer = combineReducers({
+  value : userReducer
+})
+
+
+export const store = configureStore({
+  reducer: rootReducer
+    
+})
+
+
+export type RootState = ReturnType<typeof store.getState>
+
+export type AppDispatch = typeof store.dispatch
